Add a nested route for botwizard
